requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
psycopg[binary]
anthropic
python-dotenv
Pillow
cryptography
pytest
